intelishop_backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
gunicorn
whitenoise
python-dotenv
openai
faiss-cpu
google-generativeai
edge-tts
pillow
django-import-export
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
cloudinary
django-cloudinary-storage
django-allauth
dj-rest-auth